+++ src/requisition-initiate/requisition-initiate.controller.js
@@ -144,7 +144,7 @@
                 .then(function() {
                     requisitionService.initiate(vm.facility.id, vm.program.id, selectedPeriod.id, vm.emergency, key)
                         .then(function(data) {
-                            vm.goToRequisition(data.id);
+                            goToInitiatedRequisition(data);
                         })
                         .catch(function() {
                             notificationService.error('requisitionInitiate.couldNotInitiateRequisition');
@@ -190,6 +190,13 @@
         function goToRequisition(id) {
             $state.go('openlmis.requisitions.requisition.fullSupply', {
                 rnr: id
+            });
+        }
+
+        function goToInitiatedRequisition(requisition) {
+            $state.go('openlmis.requisitions.requisition.fullSupply', {
+                rnr: requisition.id,
+                requisition: requisition
             });
         }
     }